Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
bcrypt
python-jose[cryptography]
pyjwt[crypto]          # This is the modern replacement
python-multipart